refactor: changing paths to new js files
