+++ apps/server/src/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const app = express();
-const mainRoutes = require('./router/v1/mainRoutes');
+const mainRoutes = require('./router/v1/main.routes');
 const PORT = process.env.PORT || 3000;
 
 app.use('/api', mainRoutes);
