+++ apps/server/src/router/v1/plant.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const plantController = require('../../controllers/plantController');
+const plantController = require('../../controllers/plant.controller');
 
 router.get('/', plantController.getPlants);
